# Conflicts:
#	src/main/java/yep/greenFire/greenfirebackend/order/domain/entity/Order.java

+++ src/main/java/yep/greenFire/greenfirebackend/admin/notice/dto/request/NoticeCreateRequest.java
@@ -15,6 +15,6 @@ public class NoticeCreateRequest {
     @NotBlank
     private final String noticeFixable;
 
-    private final Integer noticeWriter;
+    private final Long noticeWriter;
 
 }
